fix(utils): centralize TAG_BANNED_CHARS from specs
